refactor(emdash): separate concerns for better organization
