+++ PACKAGE_ARCHITECTURE.md
@@ -138,12 +138,16 @@ admin UI. Git-owned catalog normalization, formatting, lookup, and translation
 remain in core; EmDash layers server overrides on the normalized dictionaries.
 
 Start at [`packages/emdash/src/index.ts`](./packages/emdash/src/index.ts). The
-route boundary lives in [`packages/emdash/src/routes.ts`](./packages/emdash/src/routes.ts),
+route boundary lives in
+[`packages/emdash/src/server/routes/routes.ts`](./packages/emdash/src/server/routes/routes.ts),
 the native UI in [`packages/emdash/src/admin.tsx`](./packages/emdash/src/admin.tsx),
 the catalog override model in
 [`packages/emdash/src/catalog.ts`](./packages/emdash/src/catalog.ts), and the
 companion Astro integration in
-[`packages/emdash/src/astro.ts`](./packages/emdash/src/astro.ts). EmDash depends
+[`packages/emdash/src/astro.ts`](./packages/emdash/src/astro.ts). Server-only
+code lives under [`packages/emdash/src/server/`](./packages/emdash/src/server/)
+and the Astro runtime under
+[`packages/emdash/src/runtime/`](./packages/emdash/src/runtime/). EmDash depends
 on providers for Workers AI and CLI for the catalog commands exposed by its
 binary.
