+++ packages/emdash/src/admin/components/ErrorMessage/ErrorMessage.tsx
@@ -1,0 +1,6 @@
+import { Banner } from "@cloudflare/kumo";
+import type { ReactNode } from "react";
+
+export function ErrorMessage({ children }: { children: ReactNode }): ReactNode {
+  return <Banner variant="error" title="PolyStella couldn't complete that request" description={children} />;
+}
