+++ packages/emdash/src/admin/styles.ts
@@ -1,0 +1,12 @@
+import type { CSSProperties } from "react";
+
+export const pageStyle: CSSProperties = { margin: "0 auto", maxWidth: 1280, padding: "32px 24px 64px" };
+export const titleStyle: CSSProperties = { fontSize: 24 };
+export const tabLabelStyle: CSSProperties = { fontSize: 16 };
+export const sectionStyle: CSSProperties = { display: "grid", gap: 20, marginTop: 24 };
+export const cardStyle: CSSProperties = { padding: 20 };
+export const stackStyle: CSSProperties = { display: "grid", gap: 12 };
+export const rowStyle: CSSProperties = { alignItems: "center", display: "flex", flexWrap: "wrap", gap: 12 };
+export const inputStyle: CSSProperties = { width: "100%" };
+export const mutedStyle: CSSProperties = { opacity: 0.7 };
+export const codeStyle: CSSProperties = { margin: 0, overflowX: "auto", whiteSpace: "pre-wrap" };
